+++ packages/react/ds/src/accordion/accordion.tsx
@@ -1,20 +1,43 @@
+import React, { ReactElement } from 'react';
+
+import { cn } from '../cn.js';
 import { AccordionItem } from './accordion-item.js';
 
-type Props = {
+export type AccordionProps = {
   children: React.ReactElement<typeof AccordionItem>[];
   iconStart?: boolean;
   dataTestid?: string;
+  variant?: 'default' | 'small';
 };
 
-export const Accordion = ({ children, iconStart, dataTestid }: Props) => {
+export const Accordion = ({
+  children,
+  iconStart,
+  dataTestid,
+  variant = 'default',
+}: AccordionProps) => {
   return (
     <div
       data-testid={dataTestid}
       data-icon-start={iconStart}
-      className="gi-max-w-prose"
       role="presentation"
     >
-      {children}
+      {React.Children.map(children, (child, index) => {
+        const isLastChild = index === React.Children.count(children) - 1;
+
+        return React.isValidElement(child) ? (
+          <div
+            className={cn('gi-border-t', {
+              'gi-border-b': isLastChild,
+            })}
+          >
+            {React.cloneElement(child as ReactElement<any>, {
+              variant,
+              iconStart,
+            })}
+          </div>
+        ) : null;
+      })}
     </div>
   );
 };